+++ scripts/generate-favicons.js
@@ -151,3 +151,8 @@ generateFavicons().catch(err => {
 
 
 
+
+
+
+
+