+++ supabase/migrations/20251225130000_ritual_flow_overhaul.sql
@@ -85,3 +85,8 @@ COMMENT ON COLUMN public.ritual_preferences.proposed_time_end IS
 
 
 
+
+
+
+
+